+++ JNF_NEAT/gene.h
@@ -10,7 +10,7 @@ struct Gene {
 		unsigned int from = 0U;
 		unsigned int to = 0U;
 		unsigned int layer = 0U;
-		double weight = 0.0;
+		float weight = 0.0;
 		const unsigned int historicalMarking = numberOfExistingGenes;
 		bool isEnabled = false;
