+++ JNF_NEAT/neural_network.h
@@ -10,7 +10,7 @@ class NeuralNetwork {
 		std::vector<Sensor> sensors;
 		std::vector<std::vector<Neuron>> hiddenLayers;
 		std::vector<Neuron *> outputNeurons;
-		std::vector<double> outputs;
+		std::vector<float> outputs;
 		bool areOutputsUpToDate = false;
 
 	public:
@@ -21,13 +21,13 @@ class NeuralNetwork {
 
 		const std::vector<Gene> & GetGenes() const;
 
-		void SetInputs(std::vector<double> & inputs);
-		const std::vector<double> & GetOrCalculateOutputs();
+		void SetInputs(std::vector<float> & inputs);
+		const std::vector<float> & GetOrCalculateOutputs();
 
 	private:
 		void GenerateOnlyEssentialGenes(unsigned int numberOfInputs, unsigned int numberOfOutputs);
 		void BuildNetworkFromGenes();
-		double GetRandomWeight() const;
+		float GetRandomWeight() const;
 		void InterpretOutputNeurons();
 		void DeleteAllNeurons() ;
 };
